pyproject.toml
[tool.poetry]
name = "chatbot_with_context"
version = "0.1.0"
description = "A chatbot that maintains conversation history and uses Google search for uncertain queries."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11.11"
openai = "^1.3"
pickle-mixin = "^1.0.2"
python-dotenv = "^1.0.1"
google-api-python-client = "^2.0.0"
googlesearch-python = "^1.1.0"

[tool.poetry.scripts]
chatbot = "chatbot_with_context:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
